A very simple C interpreter for BF.
A very simple C BF interpreter, intended to be the basis of a small number of
related BF interpreters utilising yk in different ways. This deliberately does
not do any preprocessing of the input to force the run-time to do "extra" loops
(we can precalculate `[...]` jumps, and remove non-command parts of the
program). At some point we can hopefully break those loops out into functions
and mark them as `yktrace_unroll`.