Improve the build system.
Building for yk requires a lot of flags, which the user would have to
manually specify in CFLAGS/LDFLAGS.

Since there are quite so many flags, it makes sense to build them into
the build system. With this change, the user has to only specify YK_DIR
to point the build system at the yk headers and libraries.